Updated add question swagger argument, pot PII is in a separate json
Every course in user-course-information includes all the materials, lectures and quiz attempts a user,course combination has completed. This way, we can query the user to understand if they have completed the course or not and to what percentage.

There also exists a completed field that includes the date of completion and the expiration date in case a user needs to take a course every x months (such as phishing trainings)